+++ server/src/controllers/v1/baseController.js
@@ -1,4 +1,4 @@
-import { AppError } from "../service/infrastructure/errorService.js";
+import { AppError } from "../../service/infrastructure/errorService.js";
 
 export const createCommonDependencies = (db, errorService, logger, stringService) => {
 	return {
