+++ server/src/controllers/v1/diagnosticController.js
