+++ server/src/controllers/v1/monitorController.js
@@ -11,7 +11,7 @@ import {
 	getCertificateParamValidation,
 	getHardwareDetailsByIdParamValidation,
 	getHardwareDetailsByIdQueryValidation,
-} from "../validation/joi.js";
+} from "../../validation/joi.js";
 import sslChecker from "ssl-checker";
 import { fetchMonitorCertificate } from "./controllerUtils.js";
 import BaseController from "./baseController.js";
